requirements.txt
mysql-connector-python==9.5.0
redis==7.0.1
